Moving JsonRpcMethods classes into the same package, prior to refactor (#154)
