+++ ethereum/api/src/integration-test/java/org/hyperledger/besu/ethereum/api/jsonrpc/method/EthGetBlockByHashIntegrationTest.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.ethereum.api.jsonrpc.methods;
+package org.hyperledger.besu.ethereum.api.jsonrpc.method;
 
 import static org.assertj.core.api.Assertions.assertThat;
